+++ service/src/main/kotlin/pl/szudor/auth/authority/UserAuthorityRepository.kt
@@ -1,6 +1,6 @@
-package pl.szudor.user.authority
+package pl.szudor.auth.authority
 
 import org.springframework.data.jpa.repository.JpaRepository
-import pl.szudor.user.details.UserAuthority
+import pl.szudor.auth.details.UserAuthority
 
 interface UserAuthorityRepository : JpaRepository<UserAuthority, Int>
