+++ war/src/main/kotlin/pl/szudor/auth/SecurityConfig.kt
@@ -1,4 +1,4 @@
-package pl.szudor.utils
+package pl.szudor.auth
 
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
@@ -12,7 +12,6 @@ import org.springframework.security.config.http.SessionCreationPolicy
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder
 import org.springframework.security.web.SecurityFilterChain
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter
-import pl.szudor.auth.CustomUserDetailsServiceImpl
 
 
 @Configuration
